# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.2"
  title: Simple Debts API
# during dev, should point to your local machine
host: simple-debts.herokuapp.com
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  jwt:
    description: login, get your token and put it here in format 'bearer *token*'
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /login_status:
    x-swagger-router-controller: authController
    get:
      tags:
        - login
      description: Checks user's login status
      operationId: checkLoginStatus
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
      security:
        - jwt: []
  /login/facebook:
    # binds a127 app logic to a route
    x-swagger-router-controller: authController
    get:
      tags:
        - login
      description: Logins user via facebook
      # used as the method name of the controller
      operationId: facebookLogin
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer *fb_token*
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnLogin"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Login error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /login/local:
    x-swagger-router-controller: authController
    post:
      tags:
        - login
      description: Logins user via email and password
      operationId: localLogin
      parameters:
        - in: body
          required: true
          name: credentials
          description: User's credentials
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
                format: email
              password:
                type: string
                format: password
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OnLogin"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Login error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /signup/local:
    x-swagger-router-controller: authController
    put:
      tags:
        - login
      description: Signs up and login user via email and password
      operationId: localSignUp
      parameters:
        - in: body
          required: true
          name: credentials
          description: User's credentials
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
                format: email
              password:
                type: string
                format: password
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OnLogin"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Login error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  OnLogin:
    type: object
    properties:
      user:
        type: object
        properties:
          id:
            type: string
            format: id
            description: User's Id
          name:
            type: string
          picture:
            type: string
            format: url
            description: Url of user's avatar
      token:
        type: string
        description: JWT
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string